registration with valid date + change structure + apply oop principals
