+++ pom.xml
@@ -18,6 +18,7 @@
         <maven.compiler.version>3.8.1</maven.compiler.version>
         <allure.version>2.24.0</allure.version>
         <webdrivermanager.version>5.6.2</webdrivermanager.version>
+        <slf4j.version>2.0.9</slf4j.version>
         <selenium.version>4.15.0</selenium.version>
         <selenide.version>7.0.1</selenide.version>
         <junit.version>4.13.2</junit.version>
@@ -51,6 +52,13 @@
             <groupId>io.github.bonigarcia</groupId>
             <artifactId>webdrivermanager</artifactId>
             <version>${webdrivermanager.version}</version>
+        </dependency>
+
+        <!-- https://mvnrepository.com/artifact/org.slf4j/slf4j-simple -->
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <version>${slf4j.version}</version>
         </dependency>
 
         <!-- https://mvnrepository.com/artifact/org.seleniumhq.selenium/selenium-java -->
