--- src/main/java/selenide/api/BooksData.java
@@ -1,67 +1,0 @@
-package selenide.api;
-
-import java.util.Date;
-
-//конвертируем JSON в Java на сайте POJO и копируем сюда
-public class BooksData {
-    public String isbn;
-    public String title;
-    public String subTitle;
-    public String author;
-    public Date publish_date;
-    public String publisher;
-    public Integer pages;
-    public String description;
-    public String website;
-
-    //создаём конструктор и геттеры для каждой книги
-    public BooksData(String isbn, String title, String subTitle,
-                     String author, Date publish_date, String publisher,
-                     Integer pages, String description, String website) {
-        this.isbn = isbn;
-        this.title = title;
-        this.subTitle = subTitle;
-        this.author = author;
-        this.publish_date = publish_date;
-        this.publisher = publisher;
-        this.pages = pages;
-        this.description = description;
-        this.website = website;
-    }
-
-    public String getIsbn() {
-        return isbn;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getSubTitle() {
-        return subTitle;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public Date getPublish_date() {
-        return publish_date;
-    }
-
-    public String getPublisher() {
-        return publisher;
-    }
-
-    public Integer getPages() {
-        return pages;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getWebsite() {
-        return website;
-    }
-}
