+++ src/main/java/selenide/api_module/data/Users.java
@@ -1,0 +1,4 @@
+package selenide.api_module.data;
+
+public class Users {
+}
