+++ src/main/java/selenide/api_module/driver/ChromeLauncher.java
@@ -1,0 +1,14 @@
+package selenide.api_module.driver;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+//1 Создали веб-драйвер для chrome
+public class ChromeLauncher {
+
+    public static WebDriver createDriver() {
+        WebDriverManager.chromedriver().setup();
+        return new ChromeDriver();
+    }
+}
