+++ src/main/java/selenide/api_module/pages/registerPage.java
@@ -1,0 +1,8 @@
+package selenide.api_module.pages;
+
+/**
+ *5.Элементы страницы boockStoreApplication
+ */
+
+public class registerPage {
+}
