+++ src/main/java/selenide/api_module/steps/BasicSteps.java
@@ -1,0 +1,5 @@
+package selenide.api_module.steps;
+
+public class BasicSteps {
+
+}
