--- src/main/java/selenide/driver/SafariLauncher.java
@@ -1,13 +1,0 @@
-package selenide.driver;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.safari.SafariDriver;
-
-//3 создали веб-драйвер для safari
-public class SafariLauncher {
-    public static WebDriver createDriver() {
-        WebDriverManager.safaridriver().setup();
-        return new SafariDriver();
-    }
-}
