--- src/main/java/selenide/pages/registerPage.java
@@ -1,8 +1,0 @@
-package selenide.pages;
-
-/**
- *5.Элементы страницы boockStoreApplication
- */
-
-public class registerPage {
-}
