--- src/test/java/selenideTests/ui/BookStoreTests.java
@@ -1,14 +1,0 @@
-package selenideTests.ui;
-
-import org.junit.Test;
-import selenide.pages.boockStoreApplication.BookStorePage;
-
-public class BookStoreTests {
-    private static final String URL = "https://demoqa.com/books";
-
-    @Test
-    public void checkAttributeHashMap(){
-
-        BookStorePage bookStorePage = new BookStorePage(URL);
-    }
-}
