--- src/test/java/selenideTests/ui/MainPageTests.java
@@ -1,46 +1,0 @@
-package selenideTests.ui;
-
-import org.junit.Test;
-import selenide.driver.DriverProvider;
-import selenide.pages.MainPage;
-
-/**
- *3.Тесты главной страницы
- */
-
-//DriverManager - будет открываться и закрываться каждый раз браузер
-//DriverProvider - всё будет прогоняться сплошными тестами в одном окне
-public class MainPageTests extends DriverProvider { //DriverManager или DriverProvider
-    public final static String BASE_URL = "https://demoqa.com";
-    private final MainPage mainPage = new MainPage(BASE_URL);
-
-    @Test
-    public void openElementsPage(){
-        mainPage.clickOnElementsButton();
-    }
-
-    @Test
-    public void openFormsPage(){
-        mainPage.clickOnFormsButton();
-    }
-
-    @Test
-    public void openAlertFrameWindowsPage(){
-        mainPage.clickOnAlertFrameWindowsButton();
-    }
-
-    @Test
-    public void openWidgetsPage(){
-        mainPage.clickOnWidgetsButton();
-    }
-
-    @Test
-    public void openInteractionsPage(){
-        mainPage.clickOnInteractionsButton();
-    }
-
-    @Test
-    public void openBookStoreApplicationPage(){
-        mainPage.clickOnBookStoreApplication();
-    }
-}
