+++ src/test/java/selenide_tests/ui/BookStoreTests.java
@@ -1,0 +1,16 @@
+package selenide_tests.ui;
+
+import org.junit.Test;
+
+public class BookStoreTests {
+
+    @Test
+    public void loginWithValidDate(){
+
+    }
+
+    @Test
+    public void deleteUsers(){
+
+    }
+}
