+++ src/test/java/selenium_tests/ui/LoginPageTest.java
@@ -1,0 +1,21 @@
+package selenium_tests.ui;
+
+import selenium.data.Credentials;
+import selenium.driver.BaseSeleniumTest;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.junit.Test;
+import selenium.pages.bookStoreApplication.BookStorePage;
+
+public class LoginPageTest extends BaseSeleniumTest {
+    protected static final Logger LOGGER = LogManager.getLogger();
+
+    @Test
+    public void validAuthorisation(){
+        LOGGER.info("Starting validAuthorisation test...");
+        BookStorePage bookStoreApplication = new BookStorePage();
+        bookStoreApplication.openLoginPage()
+                .auth(Credentials.TEST_USERNAME, Credentials.TEST_PASSWORD);
+        LOGGER.info("validAuthorisation test finished.");
+    }
+}
